requirements.txt
yt-dlp
tqdm
google-generativeai
lancedb
pandas
